requirements.txt
attrs==22.2.0
azure-core==1.26.2
azure-identity==1.12.0
azure-storage-blob==12.14.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
flasgger==0.9.5
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mistune==2.0.4
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
pkgutil-resolve-name==1.3.10
portalocker==2.6.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
